finished week five assignment
