+++ app/app.py
@@ -1,148 +1,32 @@
-from app import app, db, bcrypt
-from app.forms import RegisterForm, LoginForm, PostForm, ProfileUpdateForm
-from app.models import User, Post
-from flask import render_template, redirect, url_for, flash, request
-from flask_login import login_user, login_required, logout_user, current_user
-from werkzeug.utils import secure_filename
-from datetime import datetime
+from app import app
+from app.post.models import Post
+from app.auth.models import User
+from flask import render_template, request
+
+
+
+from app import login_manager
 import os
+from app.auth.auth_blueprint import auth_blueprint
+from app.post.post_blueprint import post_blueprint
+from app.user.user_blueprint import user_blueprint
+
+
+
+app.register_blueprint(auth_blueprint, url_prefix='/auth')
+app.register_blueprint(post_blueprint, url_prefix='/post')
+app.register_blueprint(user_blueprint, url_prefix='/user')
+
 
 @app.route('/')
 def index():
     posts = Post.query.all()
-    # print("All Posts Retrieved:", posts)
     return render_template('index.html', posts=posts)
 
-@app.route('/post/<int:post_id>')
-def post_detail(post_id):
-    post = Post.query.get_or_404(post_id)
-    return render_template('post_detail.html', post=post)
 
-@app.route('/register', methods=['GET', 'POST'])
-def register():
-    form = RegisterForm()
-    if form.validate_on_submit():
-        existing_user = User.query.filter_by(username=form.username.data).first()
-        existing_email = User.query.filter_by(email=form.email.data).first()
-        
-        if existing_user:
-            flash("Username already exists. Please choose a different one.", "danger")
-            return redirect(url_for('register'))
-        if existing_email:
-            flash("Email already registered. Please log in.", "danger")
-            return redirect(url_for('login'))
-         
-        hashed_pw = bcrypt.generate_password_hash(form.password.data).decode('utf-8')
-        user = User(username=form.username.data, email=form.email.data, password=hashed_pw)
-        db.session.add(user)
-        db.session.commit()
-        
-        flash("Account created successfully! You can now log in.", "success")
-        return redirect(url_for('login'))
-     
-    return render_template('register.html', form=form)
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user and bcrypt.check_password_hash(user.password, form.password.data):
-            login_user(user)
-            flash("Login successful!", "success")
-            return redirect(url_for('index'))
-        else:
-            flash("Invalid email or password", "danger")
-    return render_template('login.html', form=form)
-
-@app.route('/logout')
-@login_required
-def logout():
-    logout_user()
-    flash("You have been logged out.", "info")
-    return redirect(url_for('index'))
-
-
-@app.route('/profile', methods=['GET', 'POST'])
-@login_required
-def profile():
-    form = ProfileUpdateForm()
-    if form.validate_on_submit():
-        # Handle profile picture upload
-        if form.profile_picture.data:
-            filename = secure_filename(form.profile_picture.data.filename)
-            file_path = os.path.join('static/profile_picture', filename)
-            form.profile_picture.data.save(file_path)
-            current_user.profile_picture = filename
-
-        # Update other details
-        if form.first_name.data:
-            current_user.first_name = form.first_name.data
-        if form.last_name.data:
-            current_user.last_name = form.last_name.data
-        if form.date_of_birth.data:
-            current_user.date_of_birth = form.date_of_birth.data
-
-        db.session.commit()
-        flash("Profile updated successfully!", "success")
-        return redirect(url_for('profile'))
-    return render_template('profile.html', form=form)
-
-
-@app.route('/post/new', methods=['GET', 'POST'])
-@login_required
-def create_post():
-    form = PostForm()
-    if form.validate_on_submit():
-        image_file = None
-        if form.image.data:
-            image_file = secure_filename(form.image.data.filename)
-            image_path = os.path.join(app.root_path, 'static/post_images', image_file)
-            form.image.data.save(image_path)
-        post = Post(title=form.title.data, content=form.content.data, image_file=image_file, author=current_user)
-        db.session.add(post)
-        db.session.commit()
-        flash('Your post has been created!', 'success')
-        return redirect(url_for('index'))
-    else:
-        flash('Form validation failed. Please check your input.', 'danger')
-    return render_template('create_post.html', form=form)
-
-@app.route('/post/edit/<int:post_id>', methods=['GET', 'POST'])
-@login_required
-def edit_post(post_id):
-    post = Post.query.get_or_404(post_id)
-    if post.author != current_user:
-        flash('You are not authorized to edit this post.', 'danger')
-        return redirect(url_for('index'))
-    form = PostForm()
-    if form.validate_on_submit():
-        post.title = form.title.data
-        post.content = form.content.data
-        if form.image.data:
-            image_file = secure_filename(form.image.data.filename)
-            image_path = os.path.join(app.root_path, 'static/post_images', image_file)
-            form.image.data.save(image_path)
-            post.image_file = image_file
-        db.session.commit()
-        flash('Post updated successfully!', 'success')
-        return redirect(url_for('post_detail', post_id=post.id))
-    elif request.method == 'GET':
-        form.title.data = post.title
-        form.content.data = post.content
-    return render_template('edit_post.html', form=form, post=post)
-
-@app.route('/post/delete/<int:post_id>', methods=['POST'])
-@login_required
-def delete_post(post_id):
-    post = Post.query.get_or_404(post_id)
-    if post.author != current_user:
-        flash('You are not authorized to delete this post.', 'danger')
-        return redirect(url_for('index'))
-    db.session.delete(post)
-    db.session.commit()
-    flash('Post deleted successfully!', 'success')
-    return redirect(url_for('index'))
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
 
 
 if __name__ == '__main__':
