+++ app/auth/templates/auth/register.html
@@ -1,16 +1,20 @@
 {% extends "base.html" %}
 
-{% block title %}Login{% endblock %}
+{% block title %}Register{% endblock %}
 
 {% block content %}
 <div class="container d-flex justify-content-center">
     <div class="col-md-6">
         <div class="card shadow-sm p-4">
-            <h2 class="text-center">Login</h2>
+            <h2 class="text-center">Register</h2>
             <form method="POST">
                 {{ form.hidden_tag() }}
 
                 <div class="form-group">
+                    {{ form.username.label }}
+                    {{ form.username(class="form-control") }}
+                </div>
+                <div class="form-group mt-3">
                     {{ form.email.label }}
                     {{ form.email(class="form-control") }}
                 </div>
@@ -18,10 +22,16 @@
                     {{ form.password.label }}
                     {{ form.password(class="form-control") }}
                 </div>
-                <button type="submit" class="btn btn-primary w-100 mt-3">{{ form.submit.label }}</button>
+                <div class="form-group mt-3">
+                    {{ form.confirm_password.label }}
+                    {{ form.confirm_password(class="form-control") }}
+                </div>
+
+                <button type="submit" class="btn btn-success w-100 mt-3">{{ form.submit.label }}</button>
             </form>
+
             <p class="text-center mt-3">
-                Don't have an account? <a href="{{ url_for('register') }}">Register</a>
+                Already have an account? <a href="{{ url_for('auth.login') }}">Login</a>
             </p>
         </div>
     </div>
