--- app/models.py
@@ -1,34 +1,0 @@
-# filepath: /home/john-maxwell/Tunga-Assignments/Tunga/app/models.py
-from datetime import datetime, timezone
-from app import db
-from flask_login import UserMixin
-
-# User Model
-class User(db.Model, UserMixin):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(50), unique=True, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    password = db.Column(db.String(100), nullable=False)
-    first_name = db.Column(db.String(50), nullable=True)
-    last_name = db.Column(db.String(50), nullable=True)
-    nationality = db.Column(db.String(50), nullable=True)
-    date_of_birth = db.Column(db.Date, nullable=True)
-    profile_picture = db.Column(db.String(150), nullable=True, default='default.jpg')
-    posts = db.relationship('Post', back_populates='author', cascade="all, delete-orphan")
-
-# Blog Post Model
-class Post(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    title = db.Column(db.String(250), nullable=False)
-    content = db.Column(db.Text, nullable=False)
-    date_posted = db.Column(db.DateTime(timezone=True), nullable=False, default=lambda: datetime.now(timezone.utc))
-    image_file = db.Column(db.String(150), nullable=True)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    author = db.relationship('User', back_populates='posts')
-
-# Flask-Login User Loader
-from app import login_manager
-
-@login_manager.user_loader
-def load_user(user_id):
-    return User.query.get(int(user_id))
