+++ app/post/models.py
@@ -1,0 +1,14 @@
+from datetime import datetime, timezone
+from app import db
+# from app.auth.models import User
+# from flask_login import UserMixin
+
+
+class Post(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String(250), nullable=False)
+    content = db.Column(db.Text, nullable=False)
+    date_posted = db.Column(db.DateTime(timezone=True), nullable=False, default=lambda: datetime.now(timezone.utc))
+    image_file = db.Column(db.String(150), nullable=True)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    author = db.relationship('User', back_populates='posts')
