+++ app/templates/index.html
@@ -19,8 +19,9 @@
         </small>
         {% if current_user.is_authenticated and post.author == current_user %}
         <div class="mt-3">
-            <a href="{{ url_for('edit_post', post_id=post.id) }}" class="btn btn-warning">Edit</a>
-            <form action="{{ url_for('delete_post', post_id=post.id) }}" method="POST" style="display:inline;">
+            <a href="{{ url_for('post.edit_post', post_id=post.id) }}" class="btn btn-warning">Edit</a>
+            <form method="POST" action="{{ url_for('post.delete_post', post_id=post.id) }}">
+                <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
                 <button type="submit" class="btn btn-danger">Delete</button>
             </form>
         </div>
